// src/types/app.ts
export enum City {
  Paris = 'Paris',
  Cologne = 'Cologne',
  Brussels = 'Brussels',
  Amsterdam = 'Amsterdam',
  Hamburg = 'Hamburg',
  Dusseldorf = 'Dusseldorf'
}

export enum PlaceCardType {
  Default = 'cities',
  Wide = 'favorites',
}

export enum PlaceType {
  Apartment = 'apartment',
  Room = 'room',
  House = 'house',
  Hotel = 'hotel',
}

export interface ILocation {
  latitude: number;
  longitude: number;
  zoom: number;
}

export interface ICity {
  name: `${City}`;
  location: ILocation;
}

export interface IPlaceCard {
  id: string;
  title: string;
  type: `${PlaceType}`;
  price: number;
  city: ICity;
  location: ILocation;
  isPremium: boolean;
  isFavorite: boolean;
  rating: number;
  previewImage: string;
}

export interface IGroupedPlaceCard {
  [city: string]: IPlaceCard[];
}

export interface IHost {
  name: string;
  avatarUrl: string;
  isPro: boolean;
}

export interface IOffer extends Omit<IPlaceCard, 'previewImage'> {
  description: string;
  bedrooms: number;
  goods: string[];
  host: IHost;
  images: string[];
  maxAdults: number;
}

export interface IUser {
  name: string;
  avatarUrl: string;
  isPro: boolean;
}

export interface IReview {
  id: string;
  date: string;
  user: IUser;
  comment: string;
  rating: number;
}

export enum SortingType {
  Popular = 'Popular',
  LowToHigh = 'Price: low to high',
  HighToLow = 'Price: high to low',
  TopRated = 'Top rated first',
}

export type Status = 0 | 1;